==> Notebooks/03_market_health_scoring.ipynb <==
import pandas as pd
import numpy as np

market_features = pd.read_csv("data/processed/market_features.csv")
market_features.head()


market_features["peer_group"] = (
    market_features["tier"].astype(str) + "_" +
    market_features["category"].astype(str)
)


def percentile_score(series, higher_is_better=True):
    pct = series.rank(pct=True)
    if not higher_is_better:
        pct = 1 - pct
    return (pct * 100).clip(0, 100)


market_features["vol_score"] = (
    market_features
    .groupby("peer_group")["avg_vol_30d"]
    .transform(lambda x: percentile_score(x, higher_is_better=False))
)


market_features["drawdown_score"] = (
    market_features
    .groupby("peer_group")["max_drawdown_365d"]
    .transform(lambda x: percentile_score(x, higher_is_better=True))
)


market_features["liquidity_score"] = (
    market_features
    .groupby("peer_group")["avg_vol_mcap_ratio"]
    .transform(lambda x: percentile_score(x, higher_is_better=True))
)

# absolute liquidity floor
market_features.loc[
    market_features["avg_vol_mcap_ratio"] < 0.002,
    "liquidity_score"
] *= 0.5


market_features["trend_stability_score"] = (
    100 - (market_features["momentum_7v30"].abs() * 100)
).clip(0, 100)


market_features["market_health_raw"] = (
    0.30 * market_features["vol_score"] +
    0.30 * market_features["drawdown_score"] +
    0.25 * market_features["liquidity_score"] +
    0.15 * market_features["trend_stability_score"]
)


market_features["penalty"] = 0

market_features.loc[market_features["flag_zero_market_cap"], "penalty"] -= 30
market_features.loc[market_features["flag_zero_supply"], "penalty"] -= 30
market_features.loc[market_features["flag_missing_fdv"], "penalty"] -= 10

market_features["market_health_score"] = (
    market_features["market_health_raw"] +
    market_features["penalty"]
).clip(0, 100)


market_features.to_csv(
    "data/processed/market_health_scores.csv",
    index=False
)

market_features[
    ["token_name","tier","category","market_health_score"]
].sort_values("market_health_score", ascending=False)
