==> Notebooks/05_distribution_health_scoring.ipynb <==
import pandas as pd
import numpy as np

dist = pd.read_csv("data/processed/distribution_mvp_etherscan.csv")
dist.head()


dist["peer_group"] = dist["tier"].astype(str) + "_" + dist["category"].astype(str)


def percentile_score(series, higher_is_better=True):
    pct = series.rank(pct=True)
    if not higher_is_better:
        pct = 1 - pct
    return (pct * 100).clip(0, 100)


dist["breadth_score"] = dist.groupby("peer_group")["unique_addresses"].transform(
    lambda x: percentile_score(x, higher_is_better=True)
)

dist["sender_conc_score"] = dist.groupby("peer_group")["top10_sender_share"].transform(
    lambda x: percentile_score(x, higher_is_better=False)
)

dist["receiver_conc_score"] = dist.groupby("peer_group")["top10_receiver_share"].transform(
    lambda x: percentile_score(x, higher_is_better=False)
)

dist["verification_score"] = np.where(dist["verified_contract"], 100, 0)


dist["distribution_health_score"] = (
    0.40 * dist["breadth_score"] +
    0.20 * dist["sender_conc_score"] +
    0.25 * dist["receiver_conc_score"] +
    0.15 * dist["verification_score"]
).clip(0, 100)


dist.to_csv("data/processed/distribution_health_scores.csv", index=False)

dist[["token_name","symbol","distribution_health_score"]].sort_values(
    "distribution_health_score", ascending=False
)
